Moved Title component to shared
